Remove rest of unneeded metadata API
