+++ internal/server/singleprocess/state/application.go
@@ -56,6 +56,26 @@ func (s *State) AppGet(ref *pb.Ref_Application) (*pb.Application, error) {
 	return result, err
 }
 
+// GetFileChangeSignal checks the metadata for the given application and its
+// project, returning the value of FileChangeSignal that is most relevent.
+func (s *State) GetFileChangeSignal(scope *pb.Ref_Application) (string, error) {
+	app, err := s.AppGet(scope)
+	if err != nil {
+		return "", err
+	}
+
+	if app.FileChangeSignal != "" {
+		return app.FileChangeSignal, nil
+	}
+
+	project, err := s.ProjectGet(&pb.Ref_Project{Project: scope.Project})
+	if err != nil {
+		return "", err
+	}
+
+	return project.FileChangeSignal, nil
+}
+
 func (s *State) appPut(
 	dbTxn *bolt.Tx,
 	memTxn *memdb.Txn,
